feat(ds): Add support for tools/ subdirectory
The idea is to store flake.nix, flake.lock and envrc in a separate
tools/ directory of a repo, so that these tools don't "pollute" a repo
that otherwise has nothing to do with Nix.

To work in such a repo, you symlink those files to the project root
directory. As final touch, you gitignore those symlinks. This way git
has a clean workspace, and all files are stored nicely out of regular
mortals' sight. Perfection.