feat(ds): Add -t flag for converting to tools/ structure
The idea is, ds supports the "standard" structure of flake.nix and
flake.lock and .envrc all being in the project root.

The -t flag is a completely separate thing that just converts to a
tools/ directory, but since those 3 files still exist in the project
root as symlinks, nothing changes from the functional perspective.
It's just a commodity directory structure that allows for less sinful
committing of flake.nix, flake.lock and .direnv.