feat(secrets): Add modified action
This simplifies the logic in note and le scripts, and removes the need
for those scripts to have the dir_cksum function. Neat.

I had to work around one legitimate dir_cksum use case inside note -
that is when a new note gets created. But for that we can just use
`test -s`.