fix(ncmpcpp): Disable visualizer, outputs and clock support
What instigated this change is that ncmpcpp stopped recognizing
"spectrum" as an option. Maybe this is an error in the nix package,
because this does depends on fftw, but I believe it was enabled and it
worked before, so no idea why it's suddenly a problem.

Anyway, I realized that I never use the visualizer, so I can disable it
entirely. Which means I don't need outputs support either. And while I
was at it, I also disabled the clock, because I don't use it.