fix(nvim): Fix tty colorscheme over SSH
As far as I can tell, there is no way to reliably detect whether
(Neo)Vim is inside a truecolor terminal or not. You can check DISPLAY or
WAYLAND_DISPLAY, but this will give a false positive when connecting
remotely via SSH.

So, I just decided to go an easy obvious route and explicitly set an env
when logging into any TTY other than tty1 (which is where my graphical
session is always started). Not perfect, but more reliable and simpler.