refactor: Remove all asciidoc support
I don't use asciidoc anymore, it was only a several months endeavor 2
years ago. And I don't think it's a good format to be storing any notes
or data in. And it likely has no future.

Along with Asciidoc, I've removed the note script feature of being able
to open multiple files at once when creating a new note. This feature
was only used by the adoc implementation and I don't think it's needed.