fix: Replace fancontrol with coolercontrol
Fancontrol has some issues that I didn't feel like solving. In general,
it's a duct tape script solution that's gotten too popular.

Fortunately, CoolerControl is a more reliable tool, but it does not seem
to offer configuring through text files, only GUI. But it doesn't matter
much, because it's only a one-time setup.