feat(HM): Add ardour configuration
I tried to make it using mkOutOfStoreSymlink, but it just doesn't work
with flakes and I hated it. An activation script fits my needs much
easier.

There is one annoying imperfection -- the current config directory is
called ardour8, but that will probably change with newer versions of
Ardour. So the config will need to be maintained to reflect the latest
version. Ugh.