feat(note): Move cache outside secrets, per-host
This answers a substantial problem with the notes: they were taking too
much space! And part of the issue was storing PDFs inside the secrets.
Now, cache is stored in a separate directory, which is outside secrets.
This means caching happens at view time, not at edit time (which is
another hidden enhancement, because the simpler edits are, the lower the
possibility of errors and data loss).

I dropped the strsub 3rd argument functionality, because I genuinely
don't think it has any use anymore. I don't remember why it was
introduced.