feat(boot): Declare kernel explicitly
After adding musnix, my build seemed to use a different kernel version
for some reason. I don't know if adding the kernel explicitly fixed
anything, but in the future it's good to know if some plugin changes it
inadvertently (in which case there would be a config clash and I would
know).