fix(nvidia): Pin driver to 580.119.02 for nixbox
This is because my current GPU (GTX 1060) is too old for the latest
drivers, it's recently been dropped to legacy support. I will need to
buy a new GPU soonish...

The two middle hashes don't seem to do anything, as in their values
don't matter. But removing them introduces an error:

    Failed assertions:
    - This version of NVIDIA driver does not provide a GSP firmware.

So whatever, I'll just leave them in.